+++ kinto/core/storage/memory.py
@@ -208,7 +208,7 @@ class Storage(MemoryBasedStorage):
         # In case the timestamp was specified, the collection timestamp will
         # be different from the updated timestamp. As such, we want to return
         # the one of the record, and not the collection one.
-        if not is_specified:
+        if not is_specified or previous == current:
             current = collection_timestamp
 
         self._timestamps[collection_id][parent_id] = collection_timestamp
@@ -254,6 +254,7 @@ class Storage(MemoryBasedStorage):
         self.set_record_timestamp(collection_id, parent_id, record,
                                   modified_field=modified_field)
         self._store[collection_id][parent_id][object_id] = record
+        self._cemetery[collection_id][parent_id].pop(object_id, None)
         return record
 
     def delete(self, collection_id, parent_id, object_id,