+++ CHANGELOG.rst
@@ -32,7 +32,8 @@ Protocol is now at version **1.8**. See `API changelog <http://kinto.readthedocs
 - Fix crash when a resource is registered without record path.
 - Changed behaviour of accessible objects in permissions backend when list of
   bound permissions is empty.
-
+- Bump ``last_modified`` on record when provided value is equal to previous
+  in storage ``update()`` method (#713)
 
 3.2.0 (2016-06-14)
 ==================